reg and login complete
